# api/openapi.yaml
openapi: 3.1.0
info:
  title: Searchlight API
  version: "1"
  summary: Search and percolation over SQL-backed, memory-mapped segments.
  description: |
    Searchlight's public HTTP/JSON API (spec section 5). Requests and responses are
    JSON, except `_bulk`, which is NDJSON. Every error is problem JSON (RFC 9457) with
    a machine-readable `code` and, for a refused request, `problems` whose `loc` is a
    dotted path into the request (`query.all.2.any.0.value`, `params.wait_for_seq`,
    `line.3.upsert.id`).

    **Consistency.** Every write answers with the `seq` it committed (acknowledged =
    committed to SQL). Writes become searchable within the index's refresh interval;
    `refresh=wait_for` returns once the write is searchable on this node, and
    `refresh=true` refreshes the written shards at once. Any read takes
    `wait_for_seq=N` and waits, under the request deadline, until every change up to N
    is searchable (read-your-writes). GET of one document or saved query is realtime.

    **Limits.** Bodies over `max_body_bytes` and documents over `max_doc_bytes` are
    413, as is a `_bulk` of more than `max_bulk_ops` operations. A copy whose write
    buffer is full, or that trails its writes, answers writes with 429 and
    `Retry-After`; so does a node with `search_queue` reads in progress. Every request
    has a deadline (`request_timeout`): a search past it answers what it found with
    `timed_out`, a wait past it is 504.

...
    `/readyz` need none.
...
servers:
...
security:
  - bearer: []
...
paths:
...
  /indexes:
    get:
      tags: [indexes]
      operationId: listIndexes
      summary: Every index, by name.
      responses:
        "200":
          description: The indexes.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [indexes]
                properties:
                  indexes:
                    type: array
                    items: { $ref: "#/components/schemas/IndexInfo" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "400": { $ref: "#/components/responses/Invalid" }
        "503": { $ref: "#/components/responses/Unavailable" }
        default: { $ref: "#/components/responses/Error" }
  /indexes/{index}:
    parameters:
      - $ref: "#/components/parameters/Index"
    put:
      tags: [indexes]
      operationId: createIndex
      summary: Create an index with a mapping and settings.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                mapping: { $ref: "#/components/schemas/Mapping" }
                settings: { $ref: "#/components/schemas/IndexSettings" }
      responses:
        "201":
          description: The index.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IndexInfo" }
        "400": { $ref: "#/components/responses/Invalid" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "503": { $ref: "#/components/responses/Unavailable" }
        default: { $ref: "#/components/responses/Error" }
    get:
      tags: [indexes]
      operationId: getIndex
      summary: Describe an index.
      responses:
        "200":
          description: The index.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IndexInfo" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "400": { $ref: "#/components/responses/Invalid" }
        "503": { $ref: "#/components/responses/Unavailable" }
        default: { $ref: "#/components/responses/Error" }
    delete:
      tags: [indexes]
      operationId: deleteIndex
      summary: Drop an index with its documents and saved queries.
      responses:
        "200":
          description: Dropped.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Acknowledged" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "400": { $ref: "#/components/responses/Invalid" }
        "503": { $ref: "#/components/responses/Unavailable" }
        default: { $ref: "#/components/responses/Error" }
...
  /indexes/{index}/settings:
    parameters:
      - $ref: "#/components/parameters/Index"
    patch:
      tags: [indexes]
      operationId: patchSettings
      summary: Change the refresh interval or the replica target.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                refresh_interval: { $ref: "#/components/schemas/RefreshInterval" }
                replicas_per_shard: { type: integer, minimum: 0, maximum: 1024 }
      responses:
        "200":
          description: The index.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IndexInfo" }
        "400": { $ref: "#/components/responses/Invalid" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/Unavailable" }
        default: { $ref: "#/components/responses/Error" }
  /indexes/{index}/docs/{id}:
    parameters:
      - $ref: "#/components/parameters/Index"
      - $ref: "#/components/parameters/ID"
    put:
      tags: [documents]
      operationId: putDocument
      summary: Create or replace a document.
      parameters:
        - $ref: "#/components/parameters/IfSeq"
        - $ref: "#/components/parameters/OpType"
        - $ref: "#/components/parameters/Refresh"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DocumentBody" }
      responses:
        "200": { $ref: "#/components/responses/Written" }
        "400": { $ref: "#/components/responses/Invalid" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "413": { $ref: "#/components/responses/TooLarge" }
        "429": { $ref: "#/components/responses/TooMany" }
        "503": { $ref: "#/components/responses/Unavailable" }
        default: { $ref: "#/components/responses/Error" }
    get:
      tags: [documents]
      operationId: getDocument
      summary: Read a document (realtime).
      parameters:
        - $ref: "#/components/parameters/WaitForSeq"
      responses:
        "200":
          description: The document.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Document" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "400": { $ref: "#/components/responses/Invalid" }
        "503": { $ref: "#/components/responses/Unavailable" }
        default: { $ref: "#/components/responses/Error" }
    delete:
      tags: [documents]
      operationId: deleteDocument
      summary: Delete a document.
      parameters:
        - $ref: "#/components/parameters/IfSeq"
        - $ref: "#/components/parameters/OpType"
        - $ref: "#/components/parameters/Refresh"
      responses:
        "200": { $ref: "#/components/responses/Written" }
        "400": { $ref: "#/components/responses/Invalid" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "429": { $ref: "#/components/responses/TooMany" }
        "503": { $ref: "#/components/responses/Unavailable" }
        default: { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: A token from tokens_file; read-only tokens may only read.
  parameters:
    Index:
      name: index
      in: path
      required: true
      description: 1 to 255 bytes of lowercase letters, digits, '-', '_' and '.', starting with a letter or digit.
      schema: { type: string, pattern: "^[a-z0-9][a-z0-9._-]{0,254}$" }
    ID:
      name: id
      in: path
      required: true
      description: A document or saved-query id, 1 to 512 bytes of UTF-8 without NUL; percent-encode '/'.
      schema: { type: string, minLength: 1, maxLength: 512 }
    IfSeq:
      name: if_seq
      in: query
      description: Apply only if the target's current seq is this one (409 otherwise).
      schema: { type: integer, minimum: 1 }
    OpType:
      name: op_type
      in: query
      description: create applies only if the target does not exist (409 otherwise).
      schema: { type: string, enum: [create, upsert, index] }
    Refresh:
      name: refresh
      in: query
      description: true refreshes the written shards at once; wait_for returns once the write is searchable on this node.
      schema: { type: string, enum: ["true", wait_for, "false"] }
    WaitForSeq:
      name: wait_for_seq
      in: query
      description: Wait, under the request deadline, until every change up to this seq is searchable on the copies the read uses.
      schema: { type: integer, minimum: 0 }
  responses:
    Error:
      description: |
        Any other problem: 408 request_timeout (a body slower than read_timeout),
        413 too_large, 429 too_many_requests (backpressure, a full search_queue or
        in-flight byte budget), 499 client_closed_request, 500 internal, 503
        unavailable, 504 timeout.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Written:
      description: Committed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/WriteResponse" }
    Invalid:
      description: The request is invalid; `problems` lists where.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: No valid bearer token.
      headers:
        WWW-Authenticate:
          schema: { type: string }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Forbidden:
      description: The token is read-only.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: The index, document or saved query does not exist (code index_not_found, document_not_found, query_not_found or not_found).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Conflict:
      description: An if_seq or create condition failed (code conflict, with current_seq), or the index exists (index_exists).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    TooLarge:
      description: The body, a document or the bulk is over its limit.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    TooMany:
      description: Backpressure; retry after Retry-After seconds.
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    Unavailable:
      description: The database or a shard copy is unavailable, or the node is shutting down.
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    Acknowledged:
      type: object
      additionalProperties: false
      required: [acknowledged]
      properties:
        acknowledged: { type: boolean }
    Problem:
      type: object
      additionalProperties: false
      description: RFC 9457 problem details.
      required: [type, title, status, code]
      properties:
        type: { type: string, description: "urn:searchlight:problem:<code>" }
        title: { type: string }
        status: { type: integer }
        code:
          type: string
          enum: [invalid_request, unauthorized, forbidden, not_found, index_not_found, document_not_found, query_not_found,
            method_not_allowed, index_exists, conflict, too_large, too_many_requests, timeout, request_timeout,
            client_closed_request, unavailable, internal]
        detail: { type: string }
        request_id: { type: string }
        current_seq: { type: integer, description: A failed if_seq's current seq. }
        result: { type: string, enum: [not_found], description: A delete of a document or saved query that does not exist (404). }
        stale: { type: boolean, description: Set on a 404 read from this node's copy while the database cannot be reached. }
        problems:
          type: array
          items: { $ref: "#/components/schemas/ProblemLoc" }
    ProblemLoc:
      type: object
      additionalProperties: false
      required: [loc, message]
      properties:
        loc: { type: string, examples: ["query.all.2.any.0.value"] }
        message: { type: string }
    FieldType:
      type: string
      enum: [keyword, text, keyword_list, number, bool, date]
    Mapping:
      type: object
      additionalProperties: false
      properties:
        dynamic:
          description: true types a new field from its first value; false keeps it unindexed; strict refuses it.
          oneOf:
            - type: boolean
            - type: string
              enum: ["true", "false", strict]
        fields:
          type: object
          additionalProperties: { $ref: "#/components/schemas/FieldType" }
    RefreshInterval:
      description: A positive duration ("1s", "250ms"), or -1 to refresh only on demand.
      oneOf:
        - type: string
        - type: integer
          const: -1
    IndexSettings:
      type: object
      additionalProperties: false
      properties:
        shards: { type: integer, minimum: 1, maximum: 1024, default: 1, description: Fixed at creation. }
        replicas_per_shard: { type: integer, minimum: 0, maximum: 1024, default: 0, description: 0 means every node holds every shard. }
        refresh_interval: { $ref: "#/components/schemas/RefreshInterval" }
    IndexInfo:
      type: object
      additionalProperties: false
      required: [name, uid, version, created_at, mapping, settings, docs, queries]
      properties:
        name: { type: string }
        uid: { type: string, description: This incarnation of the index. }
        version: { type: integer }
        created_at: { type: string, format: date-time }
        mapping: { $ref: "#/components/schemas/Mapping" }
        settings: { $ref: "#/components/schemas/IndexSettings" }
        docs: { type: integer, description: Live documents searchable on this node. }
        queries: { type: integer, description: Live saved queries searchable on this node. }
    DocumentBody:
      type: object
      description: A document, at most max_doc_bytes of JSON.
    Document:
      type: object
      additionalProperties: false
      required: [id, body]
      properties:
        id: { type: string }
        seq: { type: integer, description: The seq of the change that last wrote it; pass it as if_seq. Absent on a stale read. }
        body: { $ref: "#/components/schemas/DocumentBody" }
        stale: { type: boolean, description: "The database could not be reached: read from this node's copy, which may trail the latest write." }
    WriteResponse:
      type: object
      additionalProperties: false
      required: [id, seq]
      properties:
        id: { type: string }
        seq: { type: integer, description: The seq the write committed; pass it as wait_for_seq. }
        timed_out: { type: boolean, description: The refresh wait ran out of time; the write is committed. }
...